[WIP] Rollout refactoring
