+++ ledger/sandbox-on-x/src/main/scala/com/daml/ledger/sandbox/SandboxOnXRunner.scala
@@ -34,6 +34,7 @@ import com.daml.lf.engine.{Engine, EngineConfig, ValueEnricher}
 import com.daml.logging.LoggingContext.{newLoggingContext, newLoggingContextWith}
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.metrics.{JvmMetricSet, Metrics}
+import com.daml.platform.Participant
 import com.daml.platform.akkastreams.dispatcher.Dispatcher
 import com.daml.platform.apiserver._
 import com.daml.platform.configuration.{PartyConfiguration, ServerRole}
@@ -88,23 +89,10 @@ object SandboxOnXRunner {
   private def run(
       config: Config[BridgeConfig]
   )(implicit resourceContext: ResourceContext): Resource[Unit] = {
-    implicit val actorSystem: ActorSystem = ActorSystem(RunnerName)
-    implicit val materializer: Materializer = Materializer(actorSystem)
-
     for {
-      // Take ownership of the actor system and materializer so they're cleaned up properly.
-      // This is necessary because we can't declare them as implicits in a `for` comprehension.
-      _ <- ResourceOwner.forActorSystem(() => actorSystem).acquire()
-      _ <- ResourceOwner.forMaterializer(() => materializer).acquire()
-
-      // Start the ledger
       participantConfig <- validateCombinedParticipantMode(config)
-      _ <- buildLedger(
-        config,
-        participantConfig,
-        materializer,
-        actorSystem,
-      ).acquire()
+      // Start the ledger
+      _ <- Participant.owner(config, participantConfig, RunnerName).acquire()
     } yield logInitializationHeader(config, participantConfig)
   }
 
@@ -121,321 +109,6 @@ object SandboxOnXRunner {
           new IllegalArgumentException(loggingMessage)
         }
     }
-
-  def buildLedger(implicit
-      config: Config[BridgeConfig],
-      participantConfig: ParticipantConfig,
-      materializer: Materializer,
-      actorSystem: ActorSystem,
-      metrics: Option[Metrics] = None,
-  ): ResourceOwner[(ApiServer, WriteService, IndexService)] = {
-    implicit val apiServerConfig: ApiServerConfig =
-      BridgeConfigProvider.apiServerConfig(participantConfig, config)
-    val sharedEngine = new Engine(
-      EngineConfig(
-        allowedLanguageVersions = config.allowedLanguageVersions,
-        profileDir = config.profileDir,
-        stackTraceMode = config.stackTraces,
-      )
-    )
-
-    newLoggingContextWith("participantId" -> participantConfig.participantId) {
-      implicit loggingContext =>
-        for {
-          metrics <- metrics.map(ResourceOwner.successful).getOrElse(buildMetrics)
-          translationCache = LfValueTranslationCache.Cache.newInstrumentedInstance(
-            eventConfiguration = config.lfValueTranslationEventCache,
-            contractConfiguration = config.lfValueTranslationContractCache,
-            metrics = metrics,
-          )
-
-          (stateUpdatesFeedSink, stateUpdatesSource) <- AkkaSubmissionsBridge()
-
-          servicesThreadPoolSize = Runtime.getRuntime.availableProcessors()
-          servicesExecutionContext <- buildServicesExecutionContext(
-            metrics,
-            servicesThreadPoolSize,
-          )
-
-          readServiceWithSubscriber = new BridgeReadService(
-            ledgerId = config.ledgerId,
-            maximumDeduplicationDuration = config.maxDeduplicationDuration.getOrElse(
-              BridgeConfigProvider.DefaultMaximumDeduplicationDuration
-            ),
-            stateUpdatesSource,
-          )
-
-          dbSupport <- DbSupport
-            .owner(
-              jdbcUrl = apiServerConfig.jdbcUrl,
-              serverRole = ServerRole.ApiServer,
-              connectionPoolSize = apiServerConfig.databaseConnectionPoolSize,
-              connectionTimeout = apiServerConfig.databaseConnectionTimeout,
-              metrics = metrics,
-            )
-
-          stringInterningView = createStringInterningView(apiServerConfig.jdbcUrl, metrics)
-          ledgerEndCache = MutableLedgerEndCache()
-
-          generalDispatcher <-
-            Dispatcher.owner[Offset](
-              name = "sql-ledger",
-              zeroIndex = Offset.beforeBegin,
-              headAtInitialization = Offset.beforeBegin,
-            )
-
-          updateLedgerApiLedgerEnd = (ledgerEnd: LedgerEnd) => {
-            ledgerEndCache.set((ledgerEnd.lastOffset, ledgerEnd.lastEventSeqId))
-            // the order here is very important: first we need to make data available for point-wise lookups
-            // and SQL queries, and only then we can make it available on the streams.
-            // (consider example: completion arrived on a stream, but the transaction cannot be looked up)
-            generalDispatcher.signalNewHead(ledgerEnd.lastOffset)
-          }
-
-          participantInMemoryState = ParticipantInMemoryState.build(
-            apiServerConfig = apiServerConfig,
-            metrics = metrics,
-            servicesExecutionContext = servicesExecutionContext,
-            updateLedgerApiLedgerEnd = updateLedgerApiLedgerEnd,
-          )
-
-          ledgerApiBufersUpdateFlow = LedgerBuffersUpdater.flow
-            .mapAsync(1) { batch =>
-              participantInMemoryState.updateBatch(batch)
-            }
-
-          indexerHealthChecks <- buildIndexerServer(
-            metrics,
-            new TimedReadService(readServiceWithSubscriber, metrics),
-            translationCache,
-            stringInterningView,
-            ledgerApiBufersUpdateFlow,
-          )
-
-          ledgerDao = JdbcLedgerDao.read(
-            dbSupport = dbSupport,
-            eventsPageSize = config.eventsPageSize,
-            eventsProcessingParallelism = config.eventsProcessingParallelism,
-            acsIdPageSize = config.acsIdPageSize,
-            acsIdFetchingParallelism = config.acsIdFetchingParallelism,
-            acsContractFetchingParallelism = config.acsContractFetchingParallelism,
-            acsGlobalParallelism = config.acsGlobalParallelism,
-            acsIdQueueLimit = config.acsIdQueueLimit,
-            servicesExecutionContext = servicesExecutionContext,
-            metrics = metrics,
-            lfValueTranslationCache = translationCache,
-            enricher = Some(new ValueEnricher(sharedEngine)),
-            participantId = participantConfig.participantId,
-            ledgerEndCache = ledgerEndCache,
-            stringInterning = stringInterningView,
-            materializer = materializer,
-          )
-
-          ledgerEnd <- ResourceOwner.forFuture(() => ledgerDao.lookupLedgerEnd())
-          _ = updateLedgerApiLedgerEnd(ledgerEnd)
-
-          indexService <- StandaloneIndexService(
-            ledgerId = config.ledgerId,
-            config = apiServerConfig,
-            metrics = metrics,
-            engine = sharedEngine,
-            servicesExecutionContext = servicesExecutionContext,
-            lfValueTranslationCache = translationCache,
-            generalDispatcher = generalDispatcher,
-            ledgerReadDao = ledgerDao,
-            participantInMemoryState = participantInMemoryState,
-          )
-
-          timeServiceBackend = BridgeConfigProvider.timeServiceBackend(config)
-
-          writeService <- buildWriteService(
-            stateUpdatesFeedSink,
-            indexService,
-            metrics,
-            servicesExecutionContext,
-            servicesThreadPoolSize,
-            timeServiceBackend,
-          )
-
-          apiServer <- buildStandaloneApiServer(
-            sharedEngine,
-            indexService,
-            metrics,
-            servicesExecutionContext,
-            new TimedWriteService(writeService, metrics),
-            indexerHealthChecks,
-            timeServiceBackend,
-            dbSupport,
-          )
-        } yield (apiServer, writeService, indexService)
-    }
-  }
-
-  private def createStringInterningView(jdbcUrl: String, metrics: Metrics) = {
-    val dbType = DbType.jdbcType(jdbcUrl)
-    val storageBackendFactory = StorageBackendFactory.of(dbType)
-    val stringInterningStorageBackend = storageBackendFactory.createStringInterningStorageBackend
-
-    new StringInterningView(
-      loadPrefixedEntries = (fromExclusive, toInclusive, dbDispatcher) =>
-        implicit loggingContext =>
-          dbDispatcher.executeSql(metrics.daml.index.db.loadStringInterningEntries) {
-            stringInterningStorageBackend.loadStringInterningEntries(
-              fromExclusive,
-              toInclusive,
-            )
-          }
-    )
-  }
-
-  private def buildStandaloneApiServer(
-      sharedEngine: Engine,
-      indexService: IndexService,
-      metrics: Metrics,
-      servicesExecutionContext: ExecutionContextExecutorService,
-      writeService: WriteService,
-      healthChecksWithIndexer: HealthChecks,
-      timeServiceBackend: Option[TimeServiceBackend],
-      dbSupport: DbSupport,
-  )(implicit
-      actorSystem: ActorSystem,
-      loggingContext: LoggingContext,
-      config: Config[BridgeConfig],
-      apiServerConfig: ApiServerConfig,
-  ): ResourceOwner[ApiServer] =
-    StandaloneApiServer(
-      indexService = indexService,
-      ledgerId = config.ledgerId,
-      config = apiServerConfig,
-      commandConfig = config.commandConfig,
-      partyConfig = PartyConfiguration(config.extra.implicitPartyAllocation),
-      optWriteService = Some(writeService),
-      authService = config.authService,
-      healthChecks = healthChecksWithIndexer + ("write" -> writeService),
-      metrics = metrics,
-      timeServiceBackend = timeServiceBackend,
-      otherInterceptors = BridgeConfigProvider.interceptors(config),
-      engine = sharedEngine,
-      servicesExecutionContext = servicesExecutionContext,
-      userManagementStore = PersistentUserManagementStore.cached(
-        dbSupport = dbSupport,
-        metrics = metrics,
-        cacheExpiryAfterWriteInSeconds = config.userManagementConfig.cacheExpiryAfterWriteInSeconds,
-        maxCacheSize = config.userManagementConfig.maxCacheSize,
-        maxRightsPerUser = UserManagementConfig.MaxRightsPerUser,
-        timeProvider = TimeProvider.UTC,
-      )(servicesExecutionContext, loggingContext),
-      ledgerFeatures = LedgerFeatures(
-        staticTime = timeServiceBackend.isDefined,
-        commandDeduplicationFeatures = CommandDeduplicationFeatures.of(
-          deduplicationPeriodSupport = Some(
-            CommandDeduplicationPeriodSupport.of(
-              CommandDeduplicationPeriodSupport.OffsetSupport.OFFSET_NOT_SUPPORTED,
-              CommandDeduplicationPeriodSupport.DurationSupport.DURATION_NATIVE_SUPPORT,
-            )
-          ),
-          deduplicationType = CommandDeduplicationType.ASYNC_ONLY,
-          maxDeduplicationDurationEnforced = true,
-        ),
-        contractIdFeatures = ExperimentalContractIds.of(
-          v1 = ExperimentalContractIds.ContractIdV1Support.NON_SUFFIXED
-        ),
-      ),
-      userManagementConfig = config.userManagementConfig,
-    )
-
-  private def buildIndexerServer(
-      metrics: Metrics,
-      readService: ReadService,
-      translationCache: LfValueTranslationCache.Cache,
-      stringInterningView: StringInterningView,
-      updateInMemoryBuffersFlow: Flow[(Iterable[(Offset, Update)], LedgerEndMarker), Unit, NotUsed],
-  )(implicit
-      loggingContext: LoggingContext,
-      materializer: Materializer,
-      participantConfig: ParticipantConfig,
-      config: Config[BridgeConfig],
-  ): ResourceOwner[HealthChecks] =
-    for {
-      indexerHealth <- new StandaloneIndexerServer(
-        readService = readService,
-        config = BridgeConfigProvider.indexerConfig(participantConfig, config),
-        metrics = metrics,
-        lfValueTranslationCache = translationCache,
-        stringInterningView = stringInterningView,
-        updateInMemoryBuffersFlow = updateInMemoryBuffersFlow,
-      )
-    } yield new HealthChecks(
-      "read" -> readService,
-      "indexer" -> indexerHealth,
-    )
-
-  private def buildServicesExecutionContext(
-      metrics: Metrics,
-      servicesThreadPoolSize: Int,
-  ): ResourceOwner[ExecutionContextExecutorService] =
-    ResourceOwner
-      .forExecutorService(() =>
-        new InstrumentedExecutorService(
-          Executors.newWorkStealingPool(servicesThreadPoolSize),
-          metrics.registry,
-          metrics.daml.lapi.threadpool.apiServices.toString,
-        )
-      )
-      .map(ExecutionContext.fromExecutorService)
-
-  private def buildMetrics(implicit
-      participantConfig: ParticipantConfig,
-      config: Config[BridgeConfig],
-  ): ResourceOwner[Metrics] =
-    BridgeConfigProvider
-      .createMetrics(participantConfig, config)
-      .tap(_.registry.registerAll(new JvmMetricSet))
-      .pipe { metrics =>
-        config.metricsReporter
-          .fold(ResourceOwner.unit)(reporter =>
-            ResourceOwner
-              .forCloseable(() => reporter.register(metrics.registry))
-              .map(_.start(config.metricsReportingInterval.getSeconds, TimeUnit.SECONDS))
-          )
-          .map(_ => metrics)
-      }
-
-  // Builds the write service and uploads the initialization DARs
-  private def buildWriteService(
-      feedSink: Sink[(Offset, Update), NotUsed],
-      indexService: IndexService,
-      metrics: Metrics,
-      servicesExecutionContext: ExecutionContext,
-      servicesThreadPoolSize: Int,
-      timeServiceBackend: Option[TimeServiceBackend],
-  )(implicit
-      materializer: Materializer,
-      config: Config[BridgeConfig],
-      participantConfig: ParticipantConfig,
-      loggingContext: LoggingContext,
-  ): ResourceOwner[WriteService] = {
-    implicit val ec: ExecutionContext = servicesExecutionContext
-    val bridgeMetrics = new BridgeMetrics(metrics)
-    for {
-      ledgerBridge <- LedgerBridge.owner(
-        config,
-        participantConfig,
-        indexService,
-        bridgeMetrics,
-        servicesThreadPoolSize,
-        timeServiceBackend.getOrElse(TimeProvider.UTC),
-      )
-      writeService <- ResourceOwner.forCloseable(() =>
-        new BridgeWriteService(
-          feedSink = feedSink,
-          submissionBufferSize = config.extra.submissionBufferSize,
-          ledgerBridge = ledgerBridge,
-          bridgeMetrics = bridgeMetrics,
-        )
-      )
-    } yield writeService
-  }
 
   private def logInitializationHeader(
       config: Config[BridgeConfig],
